tests for long-version

+++ src/main/java/com/laserova/calculator/service/CalculatorService.java
@@ -3,11 +3,11 @@ package com.laserova.calculator.service;
 public interface CalculatorService {
     String welcome();
 
-    int plus(int num1, int num2);
+    long plus(int num1, int num2);
 
-    int minus(int num1, int num2);
+    long minus(int num1, int num2);
 
-    int multiply(int num1, int num2);
+    long multiply(int num1, int num2);
 
     double divide(int num1, int num2);
 }
